Issue 5410/3: flower: Fix int/size conversion warnings coming from g++ 8.0.1 on x86_64.

+++ flower/include/libc-extension.hh
@@ -26,8 +26,8 @@ using namespace std;
 
 #include "config.hh"
 
-char *strnlwr (char *start, int n);
-char *strnupr (char *start, int n);
+char *strnlwr (char *start, size_t n);
+char *strnupr (char *start, size_t n);
 
 #if ! HAVE_MEMMEM               /* GNU extension. */
 void *memmem (void const *haystack, int haystack_len,
